copy solutions over as Jupyter notebooks
